Golden cookie info now works
Reworked the tooltip function to be more streamlined, and it's now called every "check" tick (every few seconds). This may be changed back to the 15 minute timer check later, as buying the upgrades is not frequent.
Update log is also now in a seperate file, for a smaller main.js file and less boring nonsense.